requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
anyio==3.0.1
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.5b1
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
flake8==3.9.2
fsspec==2021.5.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
hydra-core==1.1.1
idna==2.10
importlib-resources==5.2.2
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.7.0
jupyterlab==3.0.15
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.1
kiwisolver==1.3.1
lightning-bolts==0.3.3
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
mypy-extensions==0.4.3
nbclassic==0.2.8
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.1.1
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyDeprecate==0.3.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.3.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.5
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.1+cu111
torch-tb-profiler==0.2.0
torchaudio==0.8.1
torchmetrics==0.3.2
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.10.32
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.0.0
yamlable==1.0.3
yarl==1.6.3
zipp==3.5.0
